Open PRs as drafts for sync-python-releases (#16763)
## Summary

This is a little goofy, but it saves us a click: when automation PRs are
opened as drafts, they don't need to be cycled through closed/opened to
force the CI to run. Instead, once undrafted the CI runs.

See #16505 for an example of the closed/opened cycle hack this avoids.

## Test Plan

No functional changes besides CI automation.

Signed-off-by: William Woodruff <william@astral.sh>