Use a header rather than less prompt when paging help
This means the behaviour is uniform across pagers. Also adds a helpful
title to the man page.

This avoids hitting issues where -P causes less to exit with failure due
to lack of support. Or cases where -P just doesn't display or isn't used
(some versions of less, and any pager which doesn't get identified as
less).

Notably, this information is still not displayed when a pager is not
used but in those cases the user's terminal prompt seems sufficient.