Allow setting proxy variables via configuration
This allows users to set the HTTP, HTTPS, and no proxy variables via the configuration files like
uv.toml.

This adds ProxyUrl type to surface errors earlier to users for invalid
ProxyUrls.

Also adds a uv_toml_only variable that allows us to specify specific
options that are only intended for the uv.toml file and disinclude them
from the generated docs for pyproject.toml examples.

It also adds a new integration test for the proxy support in `uv-client`.

Signed-off-by: Eli Uriegas <eliuriegas@meta.com>